Add native LL MXFP8 dispatch (#848)
Use linear UE8M0 scale bytes per 32 elements, global token-major expert
IDs, optimized Blackwell conversion targets, and updated validation.


Copilot-Session: 59760d2a-e65f-44b6-b2e6-9c271b834d7e

+++ src/ext/ep/low_latency/config.cuh
@@ -75,7 +75,7 @@ struct DispatchDataTypeTraits<DispatchDataType::FP8_E4M3> {
 template <>
 struct DispatchDataTypeTraits<DispatchDataType::MXFP8_E4M3> {
   using ElementType = Fp8E4M3;
-  using ScaleType = float;
+  using ScaleType = uint8_t;
 };
 
 template <DispatchDataType DataType>
